Changed kernel version to 5.1.0

+++ rt-thread/components/dfs/dfs_v2/src/dfs_seq_file.c
@@ -256,7 +256,7 @@ Enomem:
     goto Done;
 }
 
-off_t dfs_seq_lseek(struct dfs_file *file, off_t offset, int whence)
+ssize_t dfs_seq_lseek(struct dfs_file *file, off_t offset, int whence)
 {
     struct dfs_seq_file *seq = file->data;
     off_t retval = -EINVAL;
